refactor: remove individual_user_template canister references
