+++ shared/rust/rust-agent/rust-agent-uniffi/src/helpers.rs
