--- shared/rust/rust-agent/rust-agent-uniffi/src/individual_user_template/mod.rs
@@ -1,2 +1,0 @@
-#![allow(dead_code, unused_imports)]
-pub mod individual_user_template_helper;
