+++ shared/rust/rust-agent/rust-agent-uniffi/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod individual_user_template;
+pub mod helpers;
 pub mod sns_ledger;
 pub mod rate_limits;
 pub mod user_post_service;
